+++ search-card/src/components/Card.js
@@ -1,6 +1,5 @@
 import React from 'react';
 
-
 const Card = ({name, email, id}) => {
     return(
         <div className='tc bg-light-green dib br3 pa3 ma2 grow bw2 shadow-5'>
