+++ search-card/src/index.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 
-import App from './App';
-import './App.css';
+import './index.css';
+import App from './containers/App';
 
 import * as serviceWorker from './serviceWorker';
 import 'tachyons';
